+++ FacturapiClient.cs
@@ -1,4 +1,5 @@
-﻿using System;
+using Facturapi.Wrappers;
+using System;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
@@ -7,15 +8,16 @@ namespace Facturapi
 {
     public class FacturapiClient : IFacturapiClient
     {
-        public Wrappers.CustomerWrapper Customer { get; private set; }
-        public Wrappers.ProductWrapper Product { get; private set; }
-        public Wrappers.InvoiceWrapper Invoice { get; private set; }
-        public Wrappers.OrganizationWrapper Organization { get; private set; }
-        public Wrappers.ReceiptWrapper Receipt { get; private set; }
-        public Wrappers.RetentionWrapper Retention { get; private set; }
-        public Wrappers.CatalogWrapper Catalog { get; private set; }
-        public Wrappers.CatalogWrapper CartaporteCatalog { get; private set; }
-        public Wrappers.ToolWrapper Tool { get; private set; }
+        public CustomerWrapper Customer { get; private set; }
+        public ProductWrapper Product { get; private set; }
+        public InvoiceWrapper Invoice { get; private set; }
+        public OrganizationWrapper Organization { get; private set; }
+        public ReceiptWrapper Receipt { get; private set; }
+        public RetentionWrapper Retention { get; private set; }
+        public CatalogWrapper Catalog { get; private set; }
+        public CatalogWrapper CartaporteCatalog { get; private set; }
+        public ToolWrapper Tool { get; private set; }
+        public WebhookWrapper Webhook { get; private set; }
         private readonly HttpClient httpClient;
 
         public FacturapiClient(string apiKey, string apiVersion = "v2")
@@ -27,15 +29,16 @@ namespace Facturapi
             };
             this.httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Basic", apiKeyBase64);
 
-            this.Customer = new Wrappers.CustomerWrapper(apiKey, apiVersion, this.httpClient);
-            this.Product = new Wrappers.ProductWrapper(apiKey, apiVersion, this.httpClient);
-            this.Invoice = new Wrappers.InvoiceWrapper(apiKey, apiVersion, this.httpClient);
-            this.Organization = new Wrappers.OrganizationWrapper(apiKey, apiVersion, this.httpClient);
-            this.Receipt = new Wrappers.ReceiptWrapper(apiKey, apiVersion, this.httpClient);
-            this.Retention = new Wrappers.RetentionWrapper(apiKey, apiVersion, this.httpClient);
-            this.Catalog = new Wrappers.CatalogWrapper(apiKey, apiVersion, this.httpClient);
-            this.CartaporteCatalog = new Wrappers.CatalogWrapper(apiKey, apiVersion, this.httpClient);
-            this.Tool = new Wrappers.ToolWrapper(apiKey, apiVersion, this.httpClient);
+            this.Customer = new CustomerWrapper(apiKey, apiVersion, this.httpClient);
+            this.Product = new ProductWrapper(apiKey, apiVersion, this.httpClient);
+            this.Invoice = new InvoiceWrapper(apiKey, apiVersion, this.httpClient);
+            this.Organization = new OrganizationWrapper(apiKey, apiVersion, this.httpClient);
+            this.Receipt = new ReceiptWrapper(apiKey, apiVersion, this.httpClient);
+            this.Retention = new RetentionWrapper(apiKey, apiVersion, this.httpClient);
+            this.Catalog = new CatalogWrapper(apiKey, apiVersion, this.httpClient);
+            this.CartaporteCatalog = new CatalogWrapper(apiKey, apiVersion, this.httpClient);
+            this.Tool = new ToolWrapper(apiKey, apiVersion, this.httpClient);
+            this.Webhook = new WebhookWrapper(apiKey, apiVersion, this.httpClient);
         }
 
         public void Dispose()