+++ Models/DomainAvailability.cs
@@ -1,0 +1,7 @@
+namespace Facturapi
+{
+    public class DomainAvailability
+    {
+        public bool Available { get; set; }
+    }
+}